Display message when user want's to create a sketch which already exists

+++ src/test/java/com/github/simkuenzi/sketchbook/TestFs.java
@@ -9,13 +9,15 @@ class TestFs {
 
     void use(Test test) throws Exception {
         Path testPath = Path.of(System.getProperty("com.github.simkuenzi.sketchbook.testfs"));
-        Files.walk(testPath).sorted(Comparator.reverseOrder()).forEach(f -> {
-            try {
-                Files.delete(f);
-            } catch (IOException e) {
-                throw new RuntimeException(e);
-            }
-        });
+        if (Files.exists(testPath)) {
+            Files.walk(testPath).sorted(Comparator.reverseOrder()).forEach(f -> {
+                try {
+                    Files.delete(f);
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        }
 
         test.run(testPath);
     }
